Change types to use full range of table in advance calculation

+++ speeduino/secondaryTables.h
@@ -1,4 +1,4 @@
 void calculateSecondaryFuel();
 bool shouldWeUseSparkTable2();
 byte getVE2();
-byte getAdvance2();
+int16_t getAdvance2();
